refactor: appendChatMessage가 ChatMessageModel 자체를 받도록 변경
